requirements.txt
Camelot==12.6.29
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.5
decorator==5.1.1
distro==1.9.0
Elixir==0.7.1
Jinja2==3.1.3
JPype1==1.5.0
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pbr==6.0.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.2.0
pycparser==2.21
pypdfium2==4.28.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.4
tabula-py==2.9.0
Tempita==0.5.2
tzdata==2024.1
xlrd==0.7.1
xlwt==0.7.2
